migrate other commands

+++ bot/cogs/other.py
@@ -14,19 +14,19 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-import contextlib
-import os
+# import contextlib
+# import os
 import random
 import string
-import typing
 from difflib import get_close_matches
 
 import aiohttp
 import discord
 import wikipedia
+from discord import app_commands
 from discord.ext import commands
 
-from bot.core import better_spellcheck, get_sciname, get_taxon, send_bird
+from bot.core import better_spellcheck, get_sciname, send_bird
 from bot.data import (
     alpha_codes,
     birdListMaster,
@@ -42,6 +42,28 @@ from bot.functions import CustomCooldown, build_id_list, cache, decrypt_chacha
 
 # Discord max message length is 2000 characters, leave some room just in case
 MAX_MESSAGE = 1900
+
+
+async def state_autocomplete(
+    _: discord.Interaction,
+    current: str,
+) -> list[app_commands.Choice[str]]:
+    return [
+        app_commands.Choice(name=state, value=state)
+        for state in states
+        if current.lower() in state.lower()
+    ][:25]
+
+
+async def taxon_autocomplete(
+    _: discord.Interaction,
+    current: str,
+) -> list[app_commands.Choice[str]]:
+    return [
+        app_commands.Choice(name=taxon, value=taxon)
+        for taxon in taxons
+        if current.lower() in taxon.lower()
+    ][:25]
 
 
 class Other(commands.Cog):
@@ -70,13 +92,15 @@ class Other(commands.Cog):
         return pages
 
     # Info - Gives call+image of 1 bird
-    @commands.command(
+    @commands.hybrid_command(
         brief="- Gives an image and call of a bird",
         help="- Gives an image and call of a bird. The bird name must come before any options.",
         usage="[bird] [options]",
         aliases=["i"],
     )
     @commands.check(CustomCooldown(5.0, bucket=commands.BucketType.user))
+    @app_commands.rename(arg="bird_and_filters")
+    @app_commands.describe(arg="The bird name must come before any options.")
     async def info(self, ctx, *, arg):
         logger.info("command: info")
         arg = arg.lower().strip()
@@ -84,7 +108,7 @@ class Other(commands.Cog):
         filters = Filter.parse(arg)
         if filters.vc:
             filters.vc = False
-            await ctx.send("**The VC filter is not allowed here!**")
+            await ctx.send("**The VC filter is not allowed here!**", ephemeral=True)
 
         options = filters.display()
         arg = arg.split(" ")
@@ -108,12 +132,19 @@ class Other(commands.Cog):
                     break
 
         if not bird:
-            await ctx.send("Bird not found. Are you sure it's on the list?")
+            await ctx.send(
+                "Bird not found. Are you sure it's on the list?", ephemeral=True
+            )
             return
 
-        delete = await ctx.send("Please wait a moment.")
+        if ctx.interaction is None:
+            delete = await ctx.send("Please wait a moment.")
+        else:
+            await ctx.typing()
+
+        output_message = ""
         if options:
-            await ctx.send(f"**Detected filters**: `{'`, `'.join(options)}`")
+            output_message += f"**Detected filters**: `{'`, `'.join(options)}`\n"
 
         an = "an" if bird.lower()[0] in ("a", "e", "i", "o", "u") else "a"
         await send_bird(
@@ -121,16 +152,17 @@ class Other(commands.Cog):
             bird,
             MediaType.IMAGE,
             filters,
-            message=f"Here's {an} *{bird.lower()}* image!",
+            message=output_message + f"Here's {an} *{bird.lower()}* image!",
         )
         await send_bird(
             ctx,
             bird,
             MediaType.SONG,
             filters,
-            message=f"Here's {an} *{bird.lower()}* song!",
+            message=output_message + f"Here's {an} *{bird.lower()}* song!",
         )
-        await delete.delete()
+        if ctx.interaction is None:
+            await delete.delete()
         return
 
     @staticmethod
@@ -153,26 +185,31 @@ class Other(commands.Cog):
         return currentBird
 
     # Asset command - gives original Macaulay Library asset from asset code
-    @commands.command(
+    @commands.hybrid_command(
         help="- Gives the original Macaulay Library asset from an asset code",
         aliases=["source", "original", "orig"],
     )
     @commands.check(CustomCooldown(5.0, bucket=commands.BucketType.user))
-    async def asset(self, ctx, *, arg):
+    @app_commands.describe(
+        code="The asset code to search for.",
+        bird="The bird name that corresponds to the asset.",
+    )
+    async def asset(self, ctx, code: str, *, bird: str):
         logger.info("command: asset")
 
-        arg = arg.strip().split(" ")
-        code = arg[0]
-        guess = " ".join(arg[1:]).lower().replace("-", " ").strip()
+        guess = bird.strip().lower().replace("-", " ").strip()
         if not guess:
-            await ctx.send("Please provide the bird name to get the original asset.")
+            await ctx.send(
+                "Please provide the bird name to get the original asset.",
+                ephemeral=True,
+            )
             return
 
         try:
             asset = str(int(decrypt_chacha(code).hex(), 16))
             currentBird = await self.bird_from_asset(asset)
-        except:
-            await ctx.send("**Invalid asset code!**")
+        except ValueError:
+            await ctx.send("**Invalid asset code!**", ephemeral=True)
             return
 
         url = f"https://www.macaulaylibrary.org/asset/{asset}/"
@@ -185,14 +222,15 @@ class Other(commands.Cog):
             or guess.upper() == alpha_code
         )
         if correct or ((await self.bot.is_owner(ctx.author)) and guess == "please"):
-            await ctx.send(f"**Here you go!** {url}")
+            await ctx.send(f"**Here you go!**\n{url}")
         else:
             await ctx.send(
-                f"**Sorry, that's not the correct bird.**\n*Please try again.*"
+                "**Sorry, that's not the correct bird.**\n*Please try again.*",
+                ephemeral=True,
             )
 
     # Filter command - lists available Macaulay Library filters and aliases
-    @commands.command(
+    @commands.hybrid_command(
         help="- Lists available Macaulay Library filters.", aliases=["filter"]
     )
     @commands.check(CustomCooldown(8.0, bucket=commands.BucketType.user))
@@ -221,11 +259,13 @@ class Other(commands.Cog):
         await ctx.send(embed=embed)
 
     # List command - argument is state/bird list
-    @commands.command(
+    @commands.hybrid_command(
         help="- DMs the user with the appropriate bird list.", name="list"
     )
     @commands.check(CustomCooldown(5.0, bucket=commands.BucketType.user))
-    async def list_of_birds(self, ctx, state: str = "blank"):
+    @app_commands.describe(state="The specific bird list.")
+    @app_commands.autocomplete(state=state_autocomplete)
+    async def list_of_birds(self, ctx, state: str = "NATS"):
         logger.info("command: list")
 
         state = state.upper()
@@ -233,9 +273,13 @@ class Other(commands.Cog):
         if state not in states:
             logger.info("invalid state")
             await ctx.send(
-                f"**Sorry, `{state}` is not a valid state.**\n*Valid States:* `{', '.join(map(str, list(states.keys())))}`"
+                f"**Sorry, `{state}` is not a valid list.**\n*Valid Lists:* `{', '.join(map(str, list(states.keys())))}`",
+                ephemeral=True,
             )
             return
+
+        if ctx.interaction is not None:
+            await ctx.typing()
 
         state_birdlist = sorted(
             build_id_list(
@@ -267,13 +311,22 @@ class Other(commands.Cog):
         )
 
     # taxons command - argument is state/bird list
-    @commands.command(
+    @commands.hybrid_command(
         help="- DMs the user with the appropriate bird list.",
         name="taxon",
         aliases=["taxons", "orders", "families", "order", "family"],
     )
     @commands.check(CustomCooldown(5.0, bucket=commands.BucketType.user))
-    async def bird_taxons(self, ctx, taxon: str = "blank", state: str = "NATS"):
+    @app_commands.describe(
+        taxon="The specific bird taxon", state="The specific bird list."
+    )
+    @app_commands.autocomplete(taxon=taxon_autocomplete, state=state_autocomplete)
+    async def bird_taxons(
+        self,
+        ctx,
+        taxon: str,
+        state: str = "NATS",
+    ):
         logger.info("command: taxons")
 
         taxon = taxon.lower()
@@ -282,16 +335,21 @@ class Other(commands.Cog):
         if taxon not in taxons:
             logger.info("invalid taxon")
             await ctx.send(
-                f"**Sorry, `{taxon}` is not a valid taxon.**\n*Valid taxons:* `{', '.join(map(str, list(taxons.keys())))}`"
+                f"**Sorry, `{taxon if taxon else 'no taxon'}` is not a valid taxon.**\n*Valid taxons:* `{', '.join(map(str, list(taxons.keys())))}`",
+                ephemeral=True,
             )
             return
 
         if state not in states:
             logger.info("invalid state")
             await ctx.send(
-                f"**Sorry, `{state}` is not a valid state.**\n*Valid States:* `{', '.join(map(str, list(states.keys())))}`"
+                f"**Sorry, `{state}` is not a valid state.**\n*Valid States:* `{', '.join(map(str, list(states.keys())))}`",
+                ephemeral=True,
             )
             return
+
+        if ctx.interaction is not None:
+            await ctx.typing()
 
         bird_list = sorted(
             build_id_list(
@@ -312,7 +370,8 @@ class Other(commands.Cog):
         if not bird_list and not song_bird_list:
             logger.info("no birds for taxon/state")
             await ctx.send(
-                "**Sorry, no birds could be found for the taxon/state combo.**\n*Please try again*"
+                "**Sorry, no birds could be found for the taxon/state combo.**\n*Please try again*",
+                ephemeral=True,
             )
             return
 
@@ -341,23 +400,28 @@ class Other(commands.Cog):
         )
 
     # Wiki command - argument is the wiki page
-    @commands.command(
+    @commands.hybrid_command(
         help="- Fetch the wikipedia page for any given argument", aliases=["wiki"]
     )
     @commands.check(CustomCooldown(5.0, bucket=commands.BucketType.user))
+    @app_commands.rename(arg="query")
+    @app_commands.describe(arg="A Wikipedia query")
     async def wikipedia(self, ctx, *, arg):
         logger.info("command: wiki")
         try:
             url = get_wiki_url(arg)
         except wikipedia.exceptions.DisambiguationError:
-            await ctx.send("Sorry, that page was not found. Try being more specific.")
+            await ctx.send(
+                "Sorry, that page was not found. Try being more specific.",
+                ephemeral=True,
+            )
         except wikipedia.exceptions.PageError:
-            await ctx.send("Sorry, that page was not found.")
+            await ctx.send("Sorry, that page was not found.", ephemeral=True)
         else:
             await ctx.send(url)
 
     # meme command - sends a random bird video/gif
-    @commands.command(help="- Sends a funny bird video!")
+    @commands.hybrid_command(help="- Sends a funny bird video!")
     @commands.check(
         CustomCooldown(180.0, disable=True, bucket=commands.BucketType.user)
     )
@@ -377,45 +441,29 @@ class Other(commands.Cog):
         await channel.send(message)
         await ctx.send("Ok, sent!")
 
-    # Test command - for testing purposes only
-    @commands.command(help="- test command", hidden=True)
-    @commands.is_owner()
-    async def cache(self, ctx):
-        logger.info("command: cache stats")
-        items = []
-        with contextlib.suppress(FileNotFoundError):
-            items += os.listdir("bot_files/cache/images/")
-        with contextlib.suppress(FileNotFoundError):
-            items += os.listdir("bot_files/cache/songs/")
-        stats = {
-            "sciname_cache": get_sciname.cache_info(),
-            "taxon_cache": get_taxon.cache_info(),
-            "num_downloaded_birds": len(items),
-        }
-        await ctx.send(f"```python\n{stats}```")
+    # # Test command - for testing purposes only
+    # @commands.command(help="- test command", hidden=True)
+    # @commands.is_owner()
+    # async def cache(self, ctx):
+    #     logger.info("command: cache stats")
+    #     items = []
+    #     with contextlib.suppress(FileNotFoundError):
+    #         items += os.listdir("bot_files/cache/images/")
+    #     with contextlib.suppress(FileNotFoundError):
+    #         items += os.listdir("bot_files/cache/songs/")
+    #     stats = {
+    #         "sciname_cache": get_sciname.cache_info(),
+    #         "taxon_cache": get_taxon.cache_info(),
+    #         "num_downloaded_birds": len(items),
+    #     }
+    #     await ctx.send(f"```python\n{stats}```")
 
-    # Test command - for testing purposes only
-    @commands.command(help="- test command", hidden=True)
-    @commands.is_owner()
-    async def error(self, ctx):
-        logger.info("command: error")
-        await ctx.send(1 / 0)
-
-    # Test command - for testing purposes only
-    @commands.command(help="- test command", hidden=True)
-    @commands.is_owner()
-    async def test(
-        self,
-        ctx,
-        *,
-        user: typing.Optional[typing.Union[discord.Member, discord.User, str]] = None,
-    ):
-        logger.info("command: test")
-        await ctx.send(
-            f"```\nMembers Intent: {self.bot.intents.members}\n"
-            + f"Message Mentions: {ctx.message.mentions}\n"
-            + f"User: {user}\nType: {type(user)}```"
-        )
+    # # Test command - for testing purposes only
+    # @commands.command(help="- test command", hidden=True)
+    # @commands.is_owner()
+    # async def error(self, ctx):
+    #     logger.info("command: error")
+    #     await ctx.send(1 / 0)
 
 
 async def setup(bot):
