+++ bot/core.py
@@ -282,9 +282,9 @@ async def send_bird(
         logger.info("choosing specific Screech Owl")
         bird = random.choice(screech_owls)
 
-    delete = await ctx.send("**Fetching.** This may take a while.")
     # trigger "typing" discord message
     if ctx.interaction is None:
+        delete = await ctx.send("**Fetching.** This may take a while.")
         await ctx.typing()
 
     try:
@@ -318,7 +318,9 @@ async def send_bird(
 
     if os.stat(filename).st_size > MAX_FILESIZE:  # another filesize check (4mb)
         await delete.delete()
-        await ctx.send("**Oops! File too large :(**\n*Please try again.*")
+        await ctx.send(
+            "**Oops! File too large :(**\n*Please try again.*", ephemeral=True
+        )
         return
 
     if media_type is MediaType.IMAGE:
@@ -334,23 +336,27 @@ async def send_bird(
         if audio_file is not None and audio_file.tag is not None:
             audio_file.tag.remove(filename)
 
+    output_message = ""
     if message is not None:
-        await ctx.send(message)
+        output_message += message
 
     if macaulay_asset_id:
-        await ctx.send(
-            "**Asset Code**: `"
+        output_message += (
+            "\n**Asset Code**: `"
             + encrypt_chacha(int(macaulay_asset_id).to_bytes(8, "big").strip(b"\x00"))
             + "`"
         )
 
     if media_type is MediaType.SONG and filters.vc:
+        await ctx.send(output_message)
         await voice_functions.play(ctx, filename)
     else:
         # change filename to avoid spoilers
         file_obj = discord.File(filename, filename=f"bird.{extension}")
-        await ctx.send(file=file_obj)
-    await delete.delete()
+        await ctx.send(output_message, file=file_obj)
+
+    if ctx.interaction is None:
+        await delete.delete()
 
 
 async def get_media(ctx, bird: str, media_type: MediaType, filters: Filter):
